[refactor] Domain models: categorise models
