+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/repository/mapper/ProductMapper.kt
@@ -8,9 +8,9 @@
 package com.rwmobi.kunigami.data.repository.mapper
 
 import com.rwmobi.kunigami.data.source.network.dto.ProductDetailsDto
-import com.rwmobi.kunigami.domain.model.Product
-import com.rwmobi.kunigami.domain.model.ProductDirection
-import com.rwmobi.kunigami.domain.model.ProductFeature
+import com.rwmobi.kunigami.domain.model.product.Product
+import com.rwmobi.kunigami.domain.model.product.ProductDirection
+import com.rwmobi.kunigami.domain.model.product.ProductFeature
 
 fun ProductDetailsDto.toProduct() = Product(
     code = code,
