+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/domain/model/product/Product.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.domain.model
+package com.rwmobi.kunigami.domain.model.product
 
 import androidx.compose.runtime.Immutable
 import kotlinx.datetime.Instant
