+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/destinations/tariffs/TariffsUIState.kt
@@ -8,7 +8,7 @@
 package com.rwmobi.kunigami.ui.destinations.tariffs
 
 import androidx.compose.runtime.Immutable
-import com.rwmobi.kunigami.domain.model.Product
+import com.rwmobi.kunigami.domain.model.product.Product
 import com.rwmobi.kunigami.ui.model.ErrorMessage
 
 @Immutable
