+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/domain/repository/FakeRestApiRepository.kt
@@ -7,13 +7,13 @@
 
 package com.rwmobi.kunigami.domain.repository
 
-import com.rwmobi.kunigami.domain.model.Account
-import com.rwmobi.kunigami.domain.model.Product
-import com.rwmobi.kunigami.domain.model.Rate
 import com.rwmobi.kunigami.domain.model.Tariff
+import com.rwmobi.kunigami.domain.model.account.Account
 import com.rwmobi.kunigami.domain.model.consumption.Consumption
 import com.rwmobi.kunigami.domain.model.consumption.ConsumptionDataGroup
 import com.rwmobi.kunigami.domain.model.consumption.ConsumptionDataOrder
+import com.rwmobi.kunigami.domain.model.product.Product
+import com.rwmobi.kunigami.domain.model.rate.Rate
 import kotlinx.datetime.Instant
 
 class FakeRestApiRepository : RestApiRepository {
